Handle incoming XMPP groupchat messages
